backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
google-genai==1.10.0
deepgram-sdk==3.10.1
httpx==0.28.1
python-multipart==0.0.20
